+++ packages/trpc/server/routers/viewer/teams/createInvite.handler.ts
@@ -1,7 +1,7 @@
 import { randomBytes } from "crypto";
 
+import { isTeamAdmin } from "@calcom/features/ee/teams/lib/queries";
 import { WEBAPP_URL } from "@calcom/lib/constants";
-import { isTeamAdmin } from "@calcom/lib/server/queries/membership";
 import { prisma } from "@calcom/prisma";
 import type { TrpcSessionUser } from "@calcom/trpc/server/types";
