+++ packages/trpc/server/routers/viewer/teams/deleteInvite.handler.ts
@@ -1,4 +1,4 @@
-import { isTeamAdmin } from "@calcom/lib/server/queries/membership";
+import { isTeamAdmin } from "@calcom/features/ee/teams/lib/queries";
 import { prisma } from "@calcom/prisma";
 import type { TrpcSessionUser } from "@calcom/trpc/server/types";
