+++ packages/trpc/server/routers/viewer/teams/getInternalNotesPresets.handler.ts
@@ -1,4 +1,4 @@
-import { isTeamMember } from "@calcom/lib/server/queries/membership";
+import { isTeamMember } from "@calcom/features/ee/teams/lib/queries";
 import { prisma } from "@calcom/prisma";
 import type { TrpcSessionUser } from "@calcom/trpc/server/types";
