+++ packages/trpc/server/routers/viewer/teams/getMemberAvailability.handler.ts
@@ -1,6 +1,6 @@
+import { isTeamMember } from "@calcom/features/ee/teams/lib/queries";
 import { enrichUserWithDelegationCredentialsIncludeServiceAccountKey } from "@calcom/lib/delegationCredential/server";
 import { getUserAvailabilityService } from "@calcom/lib/di/containers/GetUserAvailability";
-import { isTeamMember } from "@calcom/lib/server/queries/membership";
 import { MembershipRepository } from "@calcom/lib/server/repository/membership";
 import type { TrpcSessionUser } from "@calcom/trpc/server/types";
