+++ packages/trpc/server/routers/viewer/teams/publish.handler.ts
@@ -1,16 +1,16 @@
 import { purchaseTeamOrOrgSubscription } from "@calcom/features/ee/teams/lib/payments";
+import { isTeamAdmin } from "@calcom/features/ee/teams/lib/queries";
 import { IS_TEAM_BILLING_ENABLED, WEBAPP_URL } from "@calcom/lib/constants";
 import { Redirect } from "@calcom/lib/redirect";
 import { isOrganisationAdmin } from "@calcom/lib/server/queries/organisations";
-import { isTeamAdmin } from "@calcom/lib/server/queries/membership";
 import { TeamService } from "@calcom/lib/server/service/teamService";
+import { Prisma } from "@calcom/prisma/client";
 import { teamMetadataSchema } from "@calcom/prisma/zod-utils";
 
 import { TRPCError } from "@trpc/server";
 
 import type { TrpcSessionUser } from "../../../types";
 import type { TPublishInputSchema } from "./publish.schema";
-import { Prisma } from "@calcom/prisma/client";
 
 type PublishOptions = {
   ctx: {
