+++ packages/trpc/server/routers/viewer/teams/removeHostsFromEventTypes.handler.ts
@@ -1,4 +1,4 @@
-import { isTeamAdmin, isTeamOwner } from "@calcom/lib/server/queries/membership";
+import { isTeamAdmin, isTeamOwner } from "@calcom/features/ee/teams/lib/queries";
 import prisma from "@calcom/prisma";
 
 import { TRPCError } from "@trpc/server";
