+++ packages/trpc/server/routers/viewer/teams/update.handler.ts
@@ -1,9 +1,9 @@
 import { getOrgFullOrigin } from "@calcom/ee/organizations/lib/orgDomains";
+import { isTeamAdmin } from "@calcom/features/ee/teams/lib/queries";
 import { IS_TEAM_BILLING_ENABLED } from "@calcom/lib/constants";
 import type { IntervalLimit } from "@calcom/lib/intervalLimits/intervalLimitSchema";
 import { validateIntervalLimitOrder } from "@calcom/lib/intervalLimits/validateIntervalLimitOrder";
 import { uploadLogo } from "@calcom/lib/server/avatar";
-import { isTeamAdmin } from "@calcom/lib/server/queries/membership";
 import { prisma } from "@calcom/prisma";
 import type { Prisma } from "@calcom/prisma/client";
 import { RedirectType, RRTimestampBasis } from "@calcom/prisma/enums";
